iot_api/pyproject.toml
[project]
name = "iot-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.2",
    "loguru>=0.7.3",
    "pymysql>=1.1.2",
    "pyserial>=3.5",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
